backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai
requests==2.31.0
firebase-admin==6.2.0
google-cloud-firestore==2.11.1

# RAG avec FAISS
langchain==0.0.340
faiss-cpu==1.7.4
sentence-transformers==2.2.2
unstructured==0.10.30
python-multipart==0.0.6
pypdf==3.17.4
docx2txt==0.8

# Optional - pour traitement avancé
numpy==1.24.3

